% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{convertAPAtrapData}
\alias{convertAPAtrapData}
\title{Convert the scExpMa list generated by scAPAtrap to other format}
\usage{
convertAPAtrapData(traplist, oformat, ...)
}
\arguments{
\item{traplist}{A list with data.frame peaks.meta and Sparse Matrix peaks.count generated by generatescExpMa.}

\item{oformat}{A character string of df/seurat/sce/pac to generate different types of objects:
data.frame (df),  SeuratObject (Seurat),  SingleCellExperiment (sce), movAPA's PACdataset (PAC).
Generally, the output includes peaks.count (SparseMatrix) and peaks.meta (peakID/chr/start/end/strand/coord).
oformat=df may be problematic when the peak-cell matrix is very large. If oformat=df failed, please set oformat to other strings to represent the peak-cell counts matrix as Sparse Matrix.}

\item{...}{Arguments passed to other methods and/or advanced arguments.
Advanced arguments:
\describe{
\item{verbose }{If `TRUE` basic status updates will be printed along the way.}
\item{logf }{If not NULL, then it should be a character string denoting a file name. Then message will be written to `logf`.}
}}
}
\description{
Given peaks' count and annotation file, this function can filter peaks with low expression levels and generate different types of objects,
             including data.frame (df), list with data.frame peaks.meta and Sparse Matrix peaks.count (list), SeuratObject (Seurat),  SingleCellExperiment (sce), movAPA's PACdataset (PAC).
}
\examples{
\dontrun{
convertAPAtrapData(from=peakslist, to='pac')
}
}
