% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{countPeaks}
\alias{countPeaks}
\title{Calculate the expression level of each peak in each cell}
\usage{
countPeaks(umitools.path, input, outputdir, TenX = T)
}
\arguments{
\item{umitools.path}{The path of the umi_tools}

\item{input}{The final.bam file is generated by \code{\link{generateFinalBam}}}

\item{outputdir}{Output file directory}

\item{TenX}{Logical value. Is it 10x data ?}
}
\value{
Generate counts.tsv.gz file in the output directory
}
\description{
Calculate the expression level of each peak in each cell
}
\examples{
umitools.path <- '/home/aa/miniconda2/envs/umi_tools/bin/umi_tools'
input <- './data/final.bam'
outputdir <- './data'
countPeaks(umitools.path,input,outputdir,TenX=T)
}
