% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{findChrTails}
\alias{findChrTails}
\title{Find the precise polyA site on each chromosome}
\usage{
findChrTails(bamfile, chr, len)
}
\arguments{
\item{bamfile}{BAM file}

\item{chr}{chromosome}

\item{len}{Chromosome length}
}
\description{
Traverse each chromosome and find the site through the A-rich at the end. It should be noted that BAM files need to be indexed
}
\examples{
findChrTails(bamfile,1,1000000001)
}
