% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{findPeaks}
\alias{findPeaks}
\title{Perform peak calling on a BAM file produced from a scRNA-seq experiment}
\usage{
findPeaks(fullCov, Strand, L, maxwidth, cutoff = 0)
}
\arguments{
\item{fullCov}{The output of loadBpCoverages}

\item{L}{Sequence length}

\item{maxwidth}{Maximum peak width}

\item{cutoff}{The base-pair level cutoff to use}

\item{strand}{strand}
}
\description{
Perform peak calling on a BAM file produced from a scRNA-seq experiment
}
\examples{
forwardPeaks <-findPeaks(fullcovF,'+',98,1000)
reversePeaks <-findPeaks(fullcovR,'-',98,1000)
}
