% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{findPeaksByStrand}
\alias{findPeaksByStrand}
\title{Perform peak calling on a BAM file with single strand}
\usage{
findPeaksByStrand(
  bamFile,
  chrs = NULL,
  strand,
  L,
  maxwidth,
  cutoff = 10,
  ofile = NULL,
  ...
)
}
\arguments{
\item{bamFile}{a BAM file with index (.bai)}

\item{chrs}{Chromosome information. If not provided, then will try get chrs from the bamFile.}

\item{strand}{strand.}

\item{L}{Read length, used to filter peaks wider than L.}

\item{maxwidth}{Maximum peak width, used to filter peaks narrower than maxwidth. Wider peaks will be splitted to smaller ones.}

\item{cutoff}{The base level coverage cutoff to filter valid peaks with enough coverages, default is 10.}

\item{ofile}{output file with each line is one peak}

\item{...}{Arguments passed to other methods and/or advanced arguments.
Advanced arguments:
\describe{
\item{verbose }{If `TRUE` basic status updates will be printed along the way.}
\item{logf }{If not NULL, then it should be a character string denoting a file name. Then message will be written to `logf`.}
}}
}
\value{
If ofile=NULL, return a data frame with many columns but four fixed columns chr/strand/start/end/value, with each row being a peak.
Otherwise, output to file and return ofile name.
}
\description{
findPeaksByStrand calls peaks on a BAM file with single strand. This function combines two functions in previous scAPAtrap, loadBpCoverages and findPeaks.
}
\examples{
\dontrun{
forwardPeaks <-findPeaksByStrand(bamFile, chrs=NULL, strand='+', L=98, maxwidth=1000, cutoff=10)
reversePeaks <-findPeaksByStrand(bamFile, chrs=NULL, strand='-', L=98, maxwidth=1000, cutoff=10)
}
}
