% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{findTailsByPeaks}
\alias{findTailsByPeaks}
\title{Find tail positions around/within peak ranges}
\usage{
findTailsByPeaks(bamfile, peaksfile, d = 200, tailsfile = NULL, ...)
}
\arguments{
\item{bamfile}{A BAM file which needs the index file (.bai).}

\item{peaksfile}{The path of the peaks.saf file generated by \code{\link{generateSAF}}.}

\item{d}{Add a 5' and 3' margin of d nt to each peak to search tails. Default is 200.}

\item{tailsfile}{if not NULL then output to tailsfile, and return the tailsfile name.}

\item{...}{Arguments passed to other methods and/or advanced arguments.
Advanced arguments:
\describe{
\item{verbose }{If `TRUE` basic status updates will be printed along the way.}
\item{logf }{If not NULL, then it should be a character string denoting a file name. Then message will be written to `logf`.}
}}
}
\value{
A data frame recording tail positions with chr/strand/coord/count; or tailsfile name (if tailsfile not NULL).
}
\description{
Find the precise positions with polyA tails within d nt of peak ranges, given a peaksfile recording peak ranges.
             This function narrows the search range of tails to peak ranges to speed up tail searching, which is useful for large BAM file.
}
\examples{
\dontrun{
findTailsByPeaks(bamfile, peaksfile, d=500)
findTailsByPeaks(bamfile, peaksfile, d=500, tailsfile='output.tails')
}
}
