% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{findUniqueMap}
\alias{findUniqueMap}
\title{Filter the BAM file to preserve the unique comparison results}
\usage{
findUniqueMap(samtools.path, input, thread, sort = T, index = T)
}
\arguments{
\item{samtools.path}{The path of the samtools}

\item{input}{BAM file}

\item{thread}{Number of threads}

\item{sort}{Logical value, whether to sort}

\item{index}{Logical value, whether or not to build the index}
}
\value{
Output path of the calculated BAM file
}
\description{
Filter the BAM file to preserve the unique comparison results
}
\examples{
samtools.path <- '/home/aa/miniconda2/envs/nar_env/bin/samtools'
input <- './data/demo.bam'
findUniqueMap(samtools.path,input,24)
}
