% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{generateAlignBam}
\alias{generateAlignBam}
\title{Align the sequence to the reference genome}
\usage{
generateAlignBam(star.path, indexdir, input.seq, prefix, thread)
}
\arguments{
\item{star.path}{The path of the star.path}

\item{indexdir}{Directory of Reference Genome Index}

\item{prefix}{The prefix of the output file name}

\item{thread}{Number of threads}
}
\value{
The path of the bam file generated after alignment
}
\description{
Align the sequence to the reference genome
}
\examples{
input.seq <- 'SRR1610598_2.extracted.fastq'
generateAlignBam(star.path,indexdir,input.seq,'./BAM/rep1',12)
}
