% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{generateFinalBam}
\alias{generateFinalBam}
\title{Generate BAM files for umi_tools to calculate expression}
\usage{
generateFinalBam(featureCounts.path, samtools.path, input, peakfile, thread)
}
\arguments{
\item{featureCounts.path}{The path of the featureCounts}

\item{samtools.path}{The path of the samtools}

\item{input}{BAM file}

\item{peakfile}{The path of the peaks.saf file generated by \code{\link{generateSAF}}}

\item{thread}{Number of threads}
}
\value{
Generate final.bam file in current folder
}
\description{
Generate BAM files for umi_tools to calculate expression
}
\examples{
input <- './data/demo.bam'
peakfile <- './data/peaks.saf'
generateFinalBam(featureCounts.path,samtools.path,input,peakfile,24)
}
