% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{generateRefIndex}
\alias{generateRefIndex}
\title{Construct reference genome index}
\usage{
generateRefIndex(star.path, genome.fasta, indexdir, thread)
}
\arguments{
\item{star.path}{The path of the star.path}

\item{genome.fasta}{Reference genome sequence, fasta format}

\item{indexdir}{Directory of Reference Genome Index}

\item{thread}{Number of threads}
}
\value{
Directory of Reference Genome Index
}
\description{
Construct reference genome index
}
\examples{
genome.fasta <- '/home/aa/genome_data/mm10/Mus_musculus.GRCm38.dna.primary_assembly.fa'
indexdir <- './mm10_index/'
generateRefIndex(star.path,genome.fasta,indexdir,24)
}
