% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{generateSAF}
\alias{generateSAF}
\title{Generate peak annotation files for quantification}
\usage{
generateSAF(forwardPeaks, reversePeaks, outputdir)
}
\arguments{
\item{forwardPeaks}{The output of \code{\link{findPeaks}},where strand = +}

\item{reversePeaks}{The output of \code{\link{findPeaks}},where strand = -}

\item{outputdir}{Output file directory}
}
\value{
Generate peaks.saf file in the output directory
}
\description{
Generate peak annotation files for quantification
}
\examples{
forwardPeaks <-findPeaks(fullcovF,'+',98,1000,cutoff = 0)
reversePeaks <-findPeaks(fullcovR,'-',98,1000,cutoff = 0)
generateSAF(forwardPeaks,reversePeaks,'./data')
}
