% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{setTools}
\alias{setTools}
\title{Set tool names for running scAPAtrap}
\usage{
setTools(tools, check = TRUE, ...)
}
\arguments{
\item{tools}{A list providing full path of four tools: samtools, umitools, featureCounts, star.}

\item{check}{Whether to check the validity of values (default is TRUE).}

\item{...}{Arguments passed to other methods and/or advanced arguments.
Advanced arguments:
\describe{
\item{verbose }{If `TRUE` basic status updates will be printed along the way.}
\item{logf }{If not NULL, then it should be a character string denoting a file name. Then message will be written to `logf`.}
}}
}
\value{
A list with tool names as the names of the list.
}
\description{
setTools sets tool names with full path for running scAPAtrap and checks the validity.
}
\examples{
tools=list(samtools='/home/dell/miniconda3/envs/scAPA/bin/samtools',
           umitools='/home/dell/miniconda3/envs/scAPA/bin/umi_tools',
           featureCounts="/home/dell/miniconda3/envs/scAPA/bin/featureCounts",
           star='/home/dell/miniconda3/envs/scAPA/bin/STAR')
\dontrun{
tools=setTools(tools)
}
tools=setTools(tools, check=FALSE)
}
