% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{TRAPFILES}
\alias{TRAPFILES}
\title{TRAPFILES global variable}
\usage{
TRAPFILES(clear = FALSE)
}
\arguments{
\item{clear}{TRUE to clear TRAPFILES variable}
}
\value{
A data frame of files containing two columns (file, what).
}
\description{
TRAPFILES is a global variable after running scAPAtrap, containing two columns (file, what).
This variable is automatically updated when running certain steps in scAPAtrap.
This function get TRAPFILES or clear TRAPFILES, but cannot add/modify items (rows) in TRAPFILES.
}
\examples{
TRAPFILES()
countPeaks(umitools.path='umitools.path', input='xx.bam', outputdir='xxdir', TenX=TRUE, notRun=TRUE)
TRAPFILES()
TRAPFILES(clear=TRUE)
}
