% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{dedupByPos}
\alias{dedupByPos}
\title{Deduplication according to the comparison position}
\usage{
dedupByPos(umitools.path, input, TenX = T)
}
\arguments{
\item{umitools.path}{The path of the umi_tools}

\item{input}{BAM file,but need index files}

\item{TenX}{Logical value. Is it 10x data?}
}
\value{
Output path of the calculated BAM file
}
\description{
Deduplication according to the comparison position
}
\examples{
umitools.path <- '/home/aa/miniconda2/envs/umi_tools/bin/umi_tools'
input <- './data/demo.Uniq.sorted.bam'
dedupByPos(umitools.path,input)
}
