% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{extractBcAndUb}
\alias{extractBcAndUb}
\title{Extract barcode and UMI in sequence}
\usage{
extractBcAndUb(umitools.path, pattern, input.seq, whitelist)
}
\arguments{
\item{umitools.path}{The path of the umi_tools}

\item{pattern}{Mode used to represent barcode and UMI (e.g The barcode length is 8 and the UMI length is 4, then pattern=CCCCCCCCNNNN)}

\item{input.seq}{path of fastq file,must contain two paths, the fastq of the first path records barcode and umi information}

\item{whitelist}{whitelist barcode}
}
\description{
Extract barcode and UMI in sequence
}
\examples{
input.seq <- c('./fastq/SRR1610598_1.fastq','./fastq/SRR1610598_2.fastq')
pattern <- 'CCCCCCCCNNNN'
whitelist <- './fastq/whitelist.txt'
extractBcAndUb(umitools.path,pattern,input.seq,whitelist)
}
