% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{generateAlignBam}
\alias{generateAlignBam}
\title{Align the sequence to the reference genome with STAR}
\usage{
generateAlignBam(star.path, indexdir, input.seq, prefix, thread = 12, ...)
}
\arguments{
\item{star.path}{The path of the STAR.}

\item{indexdir}{Directory of reference genome index.}

\item{input.seq}{Input fastq file.}

\item{prefix}{The prefix of the output file name.}

\item{thread}{Number of CPU threads, default is 12.}

\item{...}{Arguments passed to other methods and/or advanced arguments.
Advanced arguments:
\describe{
\item{verbose }{If `TRUE` basic status updates will be printed along the way.}
\item{logf }{If not NULL, then it should be a character string denoting a file name. Then message will be written to `logf`.}
\item{notRun }{Default is FALSE. If `TRUE`, the Shell commands inside this function will not run but output the commad line.}
}}
}
\description{
Align the sequence to the reference genome with STAR
}
\examples{
star.path='/home/dell/miniconda3/envs/scAPA/bin/STAR'
indexdir <- "./"
input.seq <- 'SRR1610598_2.extracted.fastq'
\dontrun{
generateAlignBam(star.path, indexdir, input.seq, prefix='./BAM/rep1',12)
}
generateAlignBam(star.path, indexdir, input.seq, prefix='./BAM/rep1', notRun=TRUE)
}
