% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{generateFinalBam}
\alias{generateFinalBam}
\title{Generate a BAM file for called peaks with featureCounts and samtools}
\usage{
generateFinalBam(
  featureCounts.path,
  samtools.path,
  input,
  peakfile,
  thread = 12,
  ...
)
}
\arguments{
\item{featureCounts.path}{The path of the featureCounts tool.}

\item{samtools.path}{The path of the samtools.}

\item{input}{BAM file.}

\item{peakfile}{The path of the peaks.saf file generated by \code{\link{generateSAF}}.}

\item{thread}{Number of CPU threads, default is 12.}

\item{...}{Arguments passed to other methods and/or advanced arguments.
Advanced arguments:
\describe{
\item{verbose }{If `TRUE` basic status updates will be printed along the way.}
\item{logf }{If not NULL, then it should be a character string denoting a file name. Then message will be written to `logf`.}
}}
}
\value{
Generate final.bam file (final.bam) in the folder of the <input> dir.
}
\description{
Generate a BAM file for called peaks with featureCounts and samtools, which can be used by umi_tools to quantify peaks.
}
\examples{
\dontrun{
input <- './data/demo.bam'
peakfile <- './data/peaks.saf'
generateFinalBam(featureCounts.path, samtools.path, input, peakfile, 24)
}
}
