% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{generatescExpMa}
\alias{generatescExpMa}
\title{Generate single cell expression matrix}
\usage{
generatescExpMa(
  countsfile,
  peaksfile,
  barcode,
  tails,
  d,
  min.cells = 2,
  min.count = 0
)
}
\arguments{
\item{countsfile}{The decompressed file path of counts.tsv.gz generated by \code{\link{countPeaks}}}

\item{peaksfile}{The path of the peaks.saf file generated by \code{\link{generateSAF}}}

\item{barcode}{Vector, and remove the numbers such as -1 in barcode}

\item{tails}{The result produced by \code{\link{findTails}} or \code{\link{findChrTails}}}

\item{min.cells}{peak is expressed in at least a few cells,the default value is 2}

\item{min.count}{peak minimum expression,the default value is 0}
}
\description{
Generate single cell expression matrix
}
\section{Warning}{

The chromosome information in tails and peaks.saf chromosome information must be consistent.
For example, tails chromosome 1 is represented as 1, and peaks.saf chromosome 1 is represented as chr1, this is not possible.
it must be 1 or chr1 at the same time.
}

\examples{
countsfile <- countPeaks(umitools.path,'./data/final.bam','./data',TenX=T)
peakfile <- generateSAF(forwardPeaks,reversePeaks,'./data')
tails <- findTails(bamfile = './data/demo.bam')
barcode <- read.delim2('./data/barcodes.tsv',header = F)
barcode <- gsub('-[0-9]','',barcode$V1)
scExpMa <- generatescExpMa(countsfile,peaksfile,barcode,tails,min.cells = 2,min.count = 0)
}
