% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{loadBpCoverages}
\alias{loadBpCoverages}
\title{Load a BAM file and calculate coverages of each position with derfinder}
\usage{
loadBpCoverages(files, chrs, ...)
}
\arguments{
\item{files}{BAM files which need the corresponding index files (.bai).}

\item{chrs}{Chromosome information.}

\item{...}{Arguments passed to other methods and/or advanced arguments.
Advanced arguments:
\describe{
\item{verbose }{If `TRUE` basic status updates will be printed along the way.}
\item{logf }{If not NULL, then it should be a character string denoting a file name. Then message will be written to `logf`.}
}}
}
\value{
A list with one element per chromosome. Each element is a DataFrame with the coverage information.
}
\description{
Load a BAM file and calculate coverages of each position with derfinder
}
\examples{
\dontrun{
chrs <- c(as.character(1:19),'X','Y')
fullcovF <- loadBpCoverages('./dedup_h2.forward.sorted.bam', chrs)
}
}
