% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAPAtrap_funlib.R
\name{separateBamBystrand}
\alias{separateBamBystrand}
\title{Separate BAM files according to strand}
\usage{
separateBamBystrand(samtools.path, input, thread)
}
\arguments{
\item{samtools.path}{The path of the samtools}

\item{input}{BAM file}

\item{thread}{Number of threads}
}
\value{
String vector, the path of the forward BAM file and the reverse BAM file generated after separating the BAM
}
\description{
Separate BAM files according to strand
}
\examples{
samtools.path <- '/home/aa/miniconda2/envs/nar_env/bin/samtools'
input <- './data/demo.Uniq.sorted.dedup.bam'
separateBamBystrand(samtools.path,input,24)
}
